Add zone that starts with the same letter matching
Technically there could be zones that encircle another zone which get undone in the masking. The process done with measuring might have to be adapted here. Another issue this raises is that Leaflet sometimes spends a while displaying the geoJSON (and turf computing the geoJSON) because it became so large due to all the zones.