Add same line matching question
It would not be a stretch to say that this question does not warrant existence. Some times train stations that aren't on the line are allegedly on the line and vice versa. A better API would improve this significantly. Thankfully hider mode prevents contradictions, but this should really be refined.