feat: Allow loading coordinates from the clipboard
As hider, it might be a bit inconvenient to have to manually input a question's coordinates.
This adds a button that parses the user's clipboard for common coordinate formats and sets the latitude and longitude values from them.
I added parsing for the coordinate formats I could find in Apple and Google Maps, but if there's any common ones that are missing I can add those too.