fix: Compress hiding zone URL param
This compresses the hiding zone URL param using deflate instead of just
b64-encoding it. The URL parameter is renamed to `?hzc`, so that old
b64-encoded `?hz` URLs continue to work. In tests, an 8256 char long
string increased to 11008 chars when b64-encoded, while decreasing to
1780 chars with deflate. The `CompressionStream` API that's being used
here is part of Baseline 2023, so it should be widely available.

Closes #116

Signed-off-by: Tobias Messner <tobias@zusor.io>